Refactor LoraWanCommandsController
